Changed ids to Long type

+++ src/main/java/com/prostaks/pazar43/security/JwtTokenProvider.java
@@ -29,7 +29,7 @@ public class JwtTokenProvider {
         Date expiryDate = new Date(now.getTime() + jwtExpirationInMs);
 
         return Jwts.builder()
-                .setSubject(Integer.toString(userPrincipal.id))
+                .setSubject(Long.toString(userPrincipal.id))
                 .setIssuedAt(new Date())
                 .setExpiration(expiryDate)
                 .signWith(SignatureAlgorithm.HS512, jwtSecret)
